This Java program calculates the frequency of a specified element in a list of integers. It utilizes the Collections.frequency() method to count the occurrences of the specified element in the list. 📊
The program initializes a list of integers and specifies an element whose frequency needs to be counted. It then utilizes the Collections.frequency() method to determine the number of occurrences of the specified element in the list. Finally, it prints out the calculated frequency. This Java program demonstrates the usage of the Collections class in Java for frequency counting. 🚀